% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetCensusSurveySummary.R
\name{GetCensusSurveySummary}
\alias{GetCensusSurveySummary}
\title{GetCensusSurveySummary()}
\usage{
GetCensusSurveySummary(SurveyName, SqlServer, Database)
}
\value{
Data frame.
}
\description{
Summarizes the animal counts and related information regarding the survey titled SurveyName.
}
\examples{
# Summarize the survey titled '2024-06-26 Mentasta Caribou Population Survey'

# Execute the query into a data frame
data = GetCensusSurveySummary('2024-06-26 Mentasta Caribou Population Survey')
head(data)
}
