% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetDatabaseConnection.R
\name{GetDatabaseConnection}
\alias{GetDatabaseConnection}
\title{GetDatabaseConnection()}
\usage{
GetDatabaseConnection()
}
\value{
An odbc database connection object.
}
\description{
This function returns database connection
}
\examples{
# Get a connection to the WRST_Caribou database
Connection = WRSTCaribou::GetDatabaseConnection(SqlServer,Database)
}
