% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetHerdDistributionMap.R
\name{GetHerdDistributionMap}
\alias{GetHerdDistributionMap}
\title{GetHerdDistributionMap()}
\usage{
GetHerdDistributionMap(SurveyName)
}
\value{
Plot
}
\description{
Returns a map of animal sightings during a survey.
}
\examples{
# Get a plot of the '2024-06-26 Mentasta Caribou Population Survey'

# Execute the query into a data frame
GetHerdDistributionMap('2024-06-26 Mentasta Caribou Population Survey')
}
