% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetCensusSurveySummary.R
\name{GetCensusSurveySummary}
\alias{GetCensusSurveySummary}
\title{GetCensusSurveySummary()
Summarizes the animal counts and related statistics for a WRST caribou survey specified by SurveyName.}
\usage{
GetCensusSurveySummary(SurveyName)
}
\value{
Data frame.
}
\description{
GetCensusSurveySummary()
Summarizes the animal counts and related statistics for a WRST caribou survey specified by SurveyName.
}
\examples{
# Get a summary of results for a caribou survey.
data = WRSTCaribou::GetCensusSurveySummary('2024-06-26 Mentasta Caribou Population Survey')
head(data)
}
