% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetDatabaseConnection.R
\name{GetDatabaseConnection}
\alias{GetDatabaseConnection}
\title{GetDatabaseConnection()}
\usage{
GetDatabaseConnection(SqlServer, Database)
}
\value{
An odbc database connection object.
}
\description{
This function returns an odbc database connection to the AK_ShallowLakes database. This function is mostly used internally by other functions in the NPSShallowLakes package, but you can use it as part of your own database queries.
}
\examples{
# Get a connection to the WRST_Caribou database
Connection = GetDatabaseConnection()
# Execute the query into a data frame
data = odbc::dbGetQuery(Connection,'SELECT Top 3 * FROM Surveys')
head(data)
}
